requirements.txt
# already install python3.6 and tensorflow-1.15.2
absl-py==0.15.0
astor==0.8.1
attrs==21.2.0
cached-property==1.5.2
certifi==2020.12.5
dataclasses==0.8
decorator==4.4.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.41.1
gviz-api==1.10.0
h5py==3.1.0
importlib-metadata==4.8.2
iniconfig==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-segmentation===0.2.0remat
Markdown==3.3.4
networkx==2.5.1
# numpy version must < 1.19.0
numpy==1.18.5
opt-einsum==3.3.0
packaging==21.3
pluggy==1.0.0
protobuf==4.0.0rc2
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
scipy==1.5.4
six==1.16.0
tensorboard==1.15.0
tensorboard-plugin-profile==2.5.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toml==0.10.2
toposort==1.7
typing_extensions==4.0.0
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
